Implement edit post pop-up.

+++ src/app/pages/cooking-post/cooking-post.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { ActivatedRoute, Router } from '@angular/router';
+import { AddPostDialogComponent } from 'src/app/components/add-post-dialog/add-post-dialog.component';
 import { CookingPost } from 'src/app/models/cooking-post.model';
 import { CookingService } from 'src/app/services/cooking.service';
 
@@ -17,6 +19,7 @@ export class CookingPostComponent implements OnInit {
   constructor(
     private route: ActivatedRoute,
     private router: Router,
+    private dialog: MatDialog,
     private cookingService: CookingService) { }
 
   async ngOnInit() {
@@ -27,6 +30,9 @@ export class CookingPostComponent implements OnInit {
   }
   
   async deletePost() {
+    if (!window.confirm('Are you sure you want to delete this post?')) {
+      return;
+    }
     this.deleteingPost = true;
     const id = this.route.snapshot.params['id'];
     await this.cookingService.deleteCookingPost(id);
@@ -34,7 +40,23 @@ export class CookingPostComponent implements OnInit {
     this.deleteingPost = false;
   }
 
-  editPost() {
+  async editPost() {
+    const dialogRef = this.dialog.open(AddPostDialogComponent, {
+      data: {
+        title: 'Edit Cooking Post',
+        isAdding: false,
+        post: this.cookingPost
+      }
+    });
+
+    const updatedCookingPost: CookingPost = await dialogRef.afterClosed().toPromise();
+    this.cookingPost = {
+      ...this.cookingPost,
+      ...updatedCookingPost,
+    }
+    this.deleteingPost = true;
+    await this.cookingService.saveCookingPost(this.cookingPost);
+    this.deleteingPost = false;
 
   }
