+++ src/app/pages/travel-post/travel-post.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { ActivatedRoute, Router } from '@angular/router';
+import { AddPostDialogComponent } from 'src/app/components/add-post-dialog/add-post-dialog.component';
 import { TravelPost } from 'src/app/models/travel-post.model';
 import { TravelService } from 'src/app/services/travel.service';
 
@@ -17,6 +19,7 @@ export class TravelPostComponent implements OnInit {
   constructor(
     private route: ActivatedRoute,
     private router: Router,
+    private dialog: MatDialog,
     private travelService: TravelService) { }
 
   async ngOnInit() {
@@ -27,6 +30,9 @@ export class TravelPostComponent implements OnInit {
   }
 
   async deletePost() {
+    if (!window.confirm('Are you sure you want to delete this post?')) {
+      return;
+    }
     this.deleteingPost = true;
     const id = this.route.snapshot.params['id'];
     await this.travelService.deleteTravelPost(id);
@@ -34,8 +40,23 @@ export class TravelPostComponent implements OnInit {
     this.deleteingPost = false;
   }
 
-  editPost() {
+  async editPost() {
+    const dialogRef = this.dialog.open(AddPostDialogComponent, {
+      data: {
+        title: 'Edit Travel Post',
+        isAdding: false,
+        post: this.travelPost
+      }
+    });
 
+    const updatedTravelPost: TravelPost = await dialogRef.afterClosed().toPromise();
+    this.travelPost = {
+      ...this.travelPost,
+      ...updatedTravelPost,
+    }
+    this.deleteingPost = true;
+    await this.travelService.saveTravelPost(this.travelPost);
+    this.deleteingPost = false;
   }
 
 }
